feat:valider un cheque de paiement par l'admin avec la gestion des exceptions

+++ src/main/java/com/smartshop/smartshop/repository/PaiementRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PaiementRepository extends JpaRepository<Paiement,Integer> {
+public interface PaiementRepository extends JpaRepository<Paiement,Long> {
     Paiement findByNumeroPaiment(String nom);
 }
